requirements.txt
mcp
pystac-client
pyyaml
mlcroissant
rasterio
gradio>=5.12.0
openai>=1.30.0
